Several fixed needed to enable backwards unsteady advection flow lines to exit region. Also moved errorcodes.h to include/vapor directory

+++ lib/flow/VTFieldLine.h
@@ -134,8 +134,8 @@ public:
 protected:
 	void releaseSeedMemory(void);
 	int euler_cauchy(TIME_DIR, TIME_DEP,float*, float);
-	int runge_kutta4(TIME_DIR, TIME_DEP, PointInfo&, float*, float);
-	int runge_kutta2(TIME_DIR, TIME_DEP, PointInfo&, float*, float);
+	int runge_kutta4(TIME_DIR, TIME_DEP, PointInfo&, double*, float);
+	int runge_kutta2(TIME_DIR, TIME_DEP, PointInfo&, double*, float);
 	int adapt_step( const VECTOR3& p2, 
 					const VECTOR3& p1, 
 					const VECTOR3& p0,
